% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{loss}
\alias{loss}
\title{Loss Function}
\usage{
loss(theta, y)
}
\arguments{
\item{theta}{Underlying parametes theta.}

\item{y}{Data sequence generated from underlying distributions.}
}
\value{
Returns a squared loss.
}
\description{
A loss function calculating the squared score of parameters and data sequence.
}
\examples{

#First generate data sequence with T=100, parameter space=[-10,10]
library(PFSMC)
data=datagenNoml(T=100,k=6,a=-10,b=10,sig=1)
y=data[[1]]; theta_t=data[[2]]
loss1=numeric(100)
for (i in 1:100)
{ loss1[i]=loss(theta_t[i],y[i]) }
loss1



}
