% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFtracking SMC-1.R
\name{PFSMC}
\alias{PFSMC}
\title{Particel Filter for Adaptive Bayes Learning}
\usage{
PFSMC(Y, eta, alpha, N, c, T)
}
\arguments{
\item{Y}{vector of sequential data sequence.}

\item{eta}{learning parameter. A positive value.}

\item{alpha}{mixing parameter.}

\item{N}{number of particles.}

\item{c}{effective sample size thershold.}

\item{T}{number of data sequence.}
}
\value{
\code{PFSMC} returns a list of effective sample size, normalized constants, predicted parameters theta and resample flags at each time.
}
\description{
A SMC version: dealing with degenercy
}
\examples{


#Generate true parameters with 5 change points
change_point=floor(201/(5+1))*c(1:5)
change_point=c(1,change_point,201)
a=-10;b=10
theta_list=a+(b-a)*runif(6)
theta_true=rep(0,201)
for(j in 1:(6))
{
theta_true[change_point[j]:change_point[j+1]]=theta_list[j]
}

#Generate normal data sequence
Y=rnorm(201,theta_true,rep(1,201))
Simulation<-PFSMC(Y=Y,eta=0.1,alpha=0.025,N=1000,c=0.5,T=201,theta_true=theta_true)
}
