% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleMultinomial.R
\name{resampleMultinomial}
\alias{resampleMultinomial}
\title{A multinomial resampling function}
\usage{
resampleMultinomial(w)
}
\arguments{
\item{w}{weights vector.}
}
\value{
Function return an index called ancestry vector, which is
drawn from a categorical distribution.
Replace samples with the ancestor.
}
\description{
To construct an unweighted sample from a weighted one.
Resample will be constructed according to the samples' weight.
Samples with higher weights will be selected more often than
samples with lower weights.
}
