% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleStratified.R
\name{resampleStratified}
\alias{resampleStratified}
\title{Stratified Resampling}
\usage{
resampleStratified(w, N)
}
\arguments{
\item{w}{Sample weights.}

\item{N}{Number of particles. Defalut is the length of w.}
}
\value{
Return the index that is chosen.
}
\description{
A resampling function. Stratified resampling is used to
reduce variance when Monte Carlo methods are implemented to estimate
population statistics. We pre-partition the (0,1] into n disjoint sets,
then draw independetly from each sub-intervals.
}
