% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampleSystematic.R
\name{resampleSystematic}
\alias{resampleSystematic}
\title{Systematic Resampling}
\usage{
resampleSystematic(w, N)
}
\arguments{
\item{w}{Sample weights}

\item{N}{Number of particles. Defalut is the length of w.}
}
\value{
index Return the chosen index of samples.
}
\description{
Similat to the stratified resampling. Difference is,
through whole resampling steps, random number is drawed only once.
Refer to \code{\link{resampleStratified}}
}
