% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFSMC.R
\name{PFSMC}
\alias{PFSMC}
\title{Sequential Monte Carlo Partical Filter}
\usage{
PFSMC(Y, eta, alpha, N, c, T, loss, resampling)
}
\arguments{
\item{Y}{a 1*T vector of sequential data sequence.}

\item{eta}{learning parameter. Determines the rate of weight updating process.}

\item{alpha}{mixing parameter. Determines the speed of model convergence and the rate of trakcing changepoints.}

\item{N}{number of particles to predict underlying densities.}

\item{c}{effective sample size thershold.}

\item{T}{number of data sequence. Time index.}
}
\value{
\code{PFSMC} returns a list of effective sample size, normalized constants, predicted parameters theta and resample flags at each time.
}
\description{
A kinetic prediction implemented
with Sequential Monte Carlo algorithm
}
\examples{


#Generate true parameters and data sequence with 5 change points
k=5; T=200; a=-10; b=10
library(PFSMC)
Data=datagenNoml(200,5,-10,10)
Y=Data[1]
theta_true=Data[2]

#Detecting changepoints using `PFSMC` funciton.

Simulation<-PFSMC(Y=Y,eta=10*T^(-1/3),alpha=k/(T-1),N=1000,c=0.5,T=200)
ESS=Simulation[1]
theta_hat=Simulation$theta_hat

#Result visulization
plot(theta_true,type="l",ylim=c(-12,12))
lines(theta_hat,col="red")
}
